Provider2: Add packages models and core
